# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for bookmark service.",
        "title": "Bookmark Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/gen-pass": {
            "get": {
                "description": "Generate a new password",
                "tags": [
                    "Password"
                ],
                "summary": "Generate a new password",
                "responses": {
                    "200": {
                        "description": "12345678",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health-check": {
            "get": {
                "description": "Check health of the service",
                "tags": [
                    "Health Check"
                ],
                "summary": "Check health of the service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.healthCheckResponse"
                        }
                    }
                }
            }
        },
        "/v1/links/redirect/{code}": {
            "get": {
                "description": "Get URL by code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Get URL",
                "parameters": [
                    {
                        "type": "string",
                        "format": "string",
                        "description": "Url code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request - invalid URL or validation error"
                    },
                    "404": {
                        "description": "URL not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/links/shorten": {
            "post": {
                "description": "Shortens a given URL and returns a shortened URL code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL Shortener"
                ],
                "summary": "Shorten URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "urlShortenRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.urlShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.urlShortenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid URL or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.healthCheckResponse": {
            "type": "object",
            "properties": {
                "instanceID": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "serviceName": {
                    "type": "string"
                }
            }
        },
        "handler.urlShortenRequest": {
            "type": "object",
            "required": [
                "exp",
                "url"
            ],
            "properties": {
                "exp": {
                    "type": "integer",
                    "minimum": 604800
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handler.urlShortenResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}